magit-current-sections: new function
Add new function `magit-current-sections' as a variant of
`magit-section-region-siblings'.  If the region isn't active it returns
a list with just one member, the current section.  This is a bit of a
kludge, and it likely will be replaced with a better approach
eventually, see #1321.

Use use of this in `magit-sequencer-read-args', `magit-am-apply-patches'
and `magit-format-patch'.