magit-checkout: remove "tracking branch confusion"
This is a backport of 9af9551470d1f10295ebd0aa5c1fac079a829e1d from
next, but where that completely removes the ability to create a local
branch whose upstream is a remote branch when using `magit-checkout',
this only removes the unnecessary complications that resulted from a
misunderstandings of what "tracking branches" are.

Similar functionality might later be restored on next.

Also see #1476.