magit-branch{,-and-checkout}: base default name on starting-point
In `magit-branch-read-args' which is used by these commands, read the
starting point first and if a remote branch is chosen use the branch
part of that as the default when reading the name of the new local
branch.

Previously we read the branch name first and then the starting-point,
which I feel is a more natural order (and also how these arguments are
passed to `git branch|checkout').  But doing it in that order did also
mean that we could not provide a default based on the starting point
when reading the branch name, because that obviously meant that we did
not know yet what the starting point was going to be.