magit-svn-get-{url,rev,ref}: new functions
Add new functions `magit-svn-get-url', `magit-svn-get-rev' and
`magit-svn-get-ref'.  `magit-svn-get-ref' actually already existed,
but it used to do what is now done by `magit-svn-get-rev'.  In turn
remove the rather complicated `magit-svn-expand-braces-in-branches',
`magit-svn-get-local-ref', `magit-svn-get-ref-info-cache' (yes there
was a cache), `magit-svn-get-ref-info' and the old implementation of
`magit-svn-get-ref'.  This shrinks the implementation by 4/5, just
saying.