magit-svn-mode-menu: deactivate as a whole
When `magit-svn-mode' but no branch that follows a Svn branch is
checked out, then deactivated the menu as a whole instead of keeping
it activated and then deactivate each child individually.

Since `magit-svn-enabled' is now used only once, isn't useful elsewhere
and has a name that is generic to communicate that, replace it with an
inline lambda.