harmonize structure of magit-{svn, stgit and topgit}
In `magit-stgit.el' rename section "Series Section" to just "Sections",
in `magit-topgit.el' do the same for "Topics Section", and
in `magit-svn.el' create that section by moving code previously found
in "Utilities", while moving the remainder of that section to the same
place where the respective sections are located in the other libraries.
Also move `magit-svn-mode-menu' below `magit-svn-mode'.