magit-refresh-goto-char: new function
Add new function `magit-goto-char' intended to be used during
refreshes only, and use it in `magit-section-goto-successor'
and `magit-rebase-process-sentinel'.

This is a reactionary change reverting the separation of church and
state, I mean `point' and `window-point'.  This is necessary because
a refresh always affects the status buffer even if it is not current.

What we are doing here is the simplest fix for the issue this is
causing; there are other, more complicated approaches possible.