magit-show-refs-head: new command
The key "y" is bound to `magit-show-refs-popup', which is one of these
popups which be default are suppressed and instead just invoke the
default action.  The default previously was `magit-show-refs' but that
command always prompts for a branch and does not allow the user to
select "HEAD", which is what we want when HEAD is detached.

So add a new command `magit-show-refs-head' which alwyas uses the
current branch, HEAD if it is detached.  When the popup is invoked
with a prefix argument the user can use `magit-show-refs' instead.