rename magit-commit-mode to magit-revision-mode
"Commit" is used both as a noun and as a verb.  To avoid confusion
rename `magit-commit-mode' to `magit-revision-mode'.  This mode
deals with existing commit objects, while `git-commit-mode' is used
to create them, using `PREFIX-commit-mode' for both would be confusing.

Rename some functions and variables accordingly, also rename
`magit-wash-commit' to `magit-diff-wash-revision', and
`magit-insert-commit-button' to `magit-diff-insert-revision-button'.

The context in which "commit" is used often is enough to tell whether
the action or state is ment , so don't rename every affected symbol,
e.g we keep using `magit-show-commit' and `magit-commit-at-point'.