magit-define-popup: better default name for arguments variable
The macro `magit-define-popup' defines one command and two variables.
The names of these symbols are based on the NAME argument.  The command
and the internal variable which holds constant data are both named just
`NAME'.  There are no restrictions imposed on what `NAME' can be but it
should be something like `PACKAGE-TOPIC-popup' or just `PACKAGE-TOPIC'.
Magit uses the first form so NAME is something `magit-TOPIC-popup'.

The second variable, whose value are the selected shell arguments, was
previously named `NAME-defaults', unless the `:variable' keyword
argument was used.

For most popups this name makes sense because the variable never hold
anything but the default shell arguments.  But for the diff popup this
didn't work well, because that variable is used for other things also.
So we used the name `magit-diff-options' in that case.

Now we change the names of all argument variables (including that of the
diff popup) to `magit-TOPIC-arguments'.  Note that there is no `-popup'
in that name; if NAME ends with "-popup" then the macro removes that
before appending "-arguments".