magit-reset*: Fix ambiguous argument
When both a revision and a file have the same name,
`magit-reset{,head,soft,hard}' would fail with the following error.

  fatal: ambiguous argument 'NAME': both revision and filename
  Use '--' to separate paths from revisions, like this:
  'git <command> [<revision>...] -- [<file>...]'

Add '--' to all commands to make it clear that the argument is a
revision. (This is already used in `magit-reset-index' and is the same
fix used in commit c5a6cc2577 before the reset functions were
rewritten).