use selected window to edit rebase sequence
Move the call to `magit-server-visit-args' from
`magit-run-git-sequencer' to `magit-rebase-async'; other
callers of the former don't need to set the visit arguments.

Call `magit-server-visit-args' without providing any values for the
arguments OTHER-WINDOW and ARGS.  Previously we attempted to provide
a value for ARGS, but ended up making OTHER-WINDOW non-nil (which we
don't want) and omitting ARGS (which didn't matter because that value
wasn't used anyway).