remove last section gap
Each section, except for the root section, begins with text which
directly belongs to it.  For sections which have children that is the
heading.  Then come the child sections and almost all text that follows
is part of these child sections.  The only exception is that after all
children there may be a single empty line, which serves to separate the
expanded section from its next sibling.  When the section is collapsed
that newline is also hidden.

This is the "newline after children" exception.  Until now there was one
other section: after the headers in the status buffer there was an empty
line separating from the following section which was not part of the
sections that wrap the headers.  It has been like this for a long time
and the reason I did not change it was that I felt that there should be
a separator between the headers and the first non-header section even
when the headers are collapsed (meaning that only the "Head" header is
showing).  I haven't really changed my mind on that but feel that the
elimination of this special case is much more important.  (We could
emulate the old look by adding a separator section after the headers,
which consisted of only a newline).