magit-section-forward: handle "\n after children" special case
There often is a newline after all children which does not belong to
the last child but to its parent.  Iff the parent section is expanded
this newline serves as separator.

When point was on that newline `magit-section-forward' used to go to
the first child instead of the next sibling.