magit-{sequencer,cherry-pick,revert}-in-progress-p: bugfix
Previously these functions assumed that a cherry-pick/revert is in
progress if ".git/{CHERRY_PICK_HEAD,REVERT_HEAD}" exists.  That does
not have to be the case.  However ".git/sequencer{,/todo} do exist
in both cases, and the content of ".git/sequencer/todo" tells us
which of the two it is.