make use of with-editor-server-window-alist
Add entries for commit message and rebase sequence buffers to the new
variable `with-editor-server-window-alist'.  Previously we arranged
for the buffer-local value of `server-window' to be set and then be
respected by `with-editor' every time a Git command that would create
such a buffer was called.  That was an ugly hack, because we used a
global variable to "pass information" to the code that would later
create and display the buffer.

Remove that hack consisting of the variable `magit-server-visit-args'
and the functions `magit-server-visit-args' and `magit-server-visit'.

`magit-server-visit' was also used to fiddle with window configuration
when showing the changes that are being committed; and to save the
commit arguments, to help `magit-diff-while-committing' later show the
correct "other meaningful diff".  Turns out this function doesn't need
our help to do so. And the window configuration tricks are now performed
in the new `magit-commit-switch-to-buffer'.