magit-refresh-buffer: wrestle with window-point again
There have been various attempts at this but they only worked in
some cases and tended to make matters worse in more cases.

The problem we have to solve is that when a non-status buffer is
refreshed and the respective status buffer exists, then that has to
be refreshed too; but unless we take extra precautions `point' would
later (when we actually display the buffer) be reset to the value of
`window-point'.  Now we update `window-point' during refresh for all
windows which do or used to display the buffer being refreshed.