magit-log-popup: add -u (for verbose logs)
Add `-u' to `magit-log-popup' and enable it by default.  This is
particularly useful when something like `-G' is also used.  When
a "oneline" log is used `-u' is removed from the arguments, the
documentation says so.

Note that the diff (like the diffstat too) is not parsed and turned
into sections like in diff buffers, but is instead inserted as is.
`magit-log-wash-log` and `magit-log-wash-log` which are used to parse
all logs are not really up to the task (not even for online logs),
they parse lines one by one, blissfully unaware of context.  Fixing
that will be quite some work.