use stashes to backup uncommitted changes
Add a new backup mechanism which uses stashes to saving uncommitted
changes, replacing the old mechanism which used patch files.  This
preserves more information and also makes it easier to restore
discarded changes.

Implement this in the new library `magit-backup.el', but most of the
heavy lifting is done in `magit-stash.el'.  Some of the functions
defined there are generalized so that they are no longer restricted
to refs/stash.

Backups are automatically created before making changes to the index or
working tree.  Where the old mechanism created a backup, the new one
does to, there might be other places where it would make sense to do so.