replace magit-current-section with magit-section-highlighted-sections
Replace the variable `magit-current-section' whose value was a single
section with `magit-section-highlighted-sections' whose value is a list
of sections.

Currently all sections are still highlighted using overlays and because
these overlays are stored in existing `magit-section-highlight-overlays',
we don't actually have to store all highlighted sections in the new
variable yet.

This commit makes it possible to use text-properties instead of overlays
to highlight diff sections, but that follows in a later commit.