make the current section part of the section selection
Until now only sections that had at least one character inside the
region were part of the section selection, and so the current section
usually wasn't part of the selection.

Commands that only know how to act on a single section simply ignore
the selection and instead act on the current section.  When the current
section wasn't considered part of the selection yet this was problematic
because users don't necessarily know which commands use the selection,
so they might select A and B, while point is on C and invoke a command
with the intention to do something with A and B but instead end up
having something done to C.  Now they would expect something to happen
with A and B, and it would only happen to B, which is much less
problematic.