improve appearance of sibling selections and hunk regions
The region can be used to select either several sibling sections or a
range of lines within a single hunk section.

When several sections are selected they are now all displayed the same
way each one of them would be displayed if it were the current section.
If the selected sections are diff file or hunk sections then their
headings are additionally displayed in a different face to make it more
obvious that multiple sections are selected.

When the region spans lines within a hunk then only these lines are
propertized using the usual distinct faces for added and remove lines
lines.  Lines inside the hunk but outside the region are propertized
using the context face, that includes lines beginning with "+" and "-".

Highlighting of the current section now works like this.  The section
is highlighted using an overlay with covers all of the section.  If
multiple sections are selected each one of them is highlighted using
a separate overlay.  Each highlighting overlay has to be added to
`magit-section-highlight-overlays'.  After each command all of these
overlays are deleted and new ones created as needed, which isn't as
wasteful as it may sound.

Highlighted hunk sections and sections that contain such sections
also use an overlay but it only covers the heading.  Hunk bodies are
propertized using text properties.  Each section which uses properties
for highlighting has to be added to `magit-section-highlighted-sections'.
These text properties are not discarded after every command; properties
are only changed for those sections which are newly or no longer
highlighted.