revert all recent changes to magit-revert-buffers
This reverts
  140e8239eccb45a5673c770d1f56c01e921ecceb
  febc832cbe12739309c41e453b36067dd4c4bc55
  68af16a2cf1d3a31e90beea80aa97f5ecc9c797e

These changes have lead to a performance regression and while some users
bothered to report it or tweet about it I have so far gotten little help
in the form of replies to my questions or reliable benchmarks.

See #1544, #1545 and #1546.

Note that this reverts changes to functionality which would be
desirable.  Generally I no longer make any functional changes to the
master branch and instead concentrate getting the next branch ready
for merge into master.  I have made an exception here and it bit me.