magit-diff-{set-context,context-p}: new functions
Add the new function `magit-diff-set-context' to do all the heavy
lifting for `magit-diff-{less,more,default}-context'.  This replaces
`magit-diff-previous-context-lines' which only determined the
current (aka previous) value.

The main advantage of this approach is that this will allow using
different variables in diff-only and also-diff buffers, which will be
implemented in a later commit.

By default Git shows three context lines, but until now we did not avoid
having the redundant argument "-U3" as a member of the arguments list.
`magit-diff-set-context' is now capable of doing that, but that also
means that `(not (member "-U0" args))' no longer is sufficient to check
whether any context lines are shown.  Add new `magit-diff-context-p' for
that purpose.