use a variety of variables to hold diff arguments
Use different variables for diff arguments depending on the mode.

In diff-only buffers (whose major-mode derives from `magit-diff-mode'),
continue to use `magit-diff-arguments'.  This _should_ then be used
to set the buffer-local value of `magit-refresh-args'.  But this is
currently not the case and won't be fixed until the next commit.

In also-diff buffers (which contain diffs but also other sections,
e.g. status buffers) we cannot use `magit-refresh-args' because
the diff sections would compete with other sections over who gets
to store their information in there.  So we use the new variable
`magit-diff-section-arguments'.

A later commit will add a new popup which actually allows changing the
value of either `magit-refresh-args' or `magit-diff-section-arguments'.

To make dealing with these complications easy add the new utility
functions `magit-diff-arguments' and `magit-diff-arguments-variable'.