respect arguments set in magit-diff-popup
Some time ago this broke and the arguments set in the popup were simply
ignored.  The last few commits improved how we deal with diff arguments,
now we also need to teach the diff commands to actually retrieve the
arguments from the popup (using `magit-current-popup-args') and then
use them when calling `git diff'.