add "-M" and "-C" to diff popup
Remove the variable `magit-diff-extra-options' which was always intended
as a kludge.  Instead make "-M" and "-C" available from the diff popup.

Do _not_ add these arguments to the popup default value or to that of
the variable `magit-diff-section-arguments'.  Note that these arguments
were _not_ previously used "by default", the were used unconditionally,
because there was no other way.  Enabling this by default isn't
desirable because it decreases performance.

Also use "--no-prefix" (which also was a member of the removed variable,
but unlike the other arguments is mandatory) explicitly where previously
the variable was used.