allow changing diff arguments used in current buffer
Add a new popup `magit-diff-refresh-popup' and a new command
`magit-diff-refresh' invoked from that popup, to allow changing the diff
arguments used in the current buffer.  In a diff buffer this changes the
value of `magit-refresh-args' and refreshes.  In other buffers which may
contain other things besides diffs change `magit-diff-section-arguments'
and refresh.

Also add `magit-diff-set-default-arguments' and
`magit-diff-save-default-arguments' which additionally set/save the
default value of the appropriate variable.  In diff buffers that is
`magit-diff-arguments', not `magit-refresh-args'.