magit-diff-dwim: new command
In the `magit-diff-popup' bind "d" to the new command, which means that
we can no longer use that for `magit-diff-unstaged'.  Instead "u" shows
unstaged changes and "s" shows changes, which by the way I intended to
do eventually every since I started using "d" and "i" for these
variants.

This also means that we have to rearrange the actions in the popup.
And that in turn is a good opportunity to also make `magit-stash-show'
available from this popup, allowing that command to take advantage
of its use of `magit-diff-arguments' in the `interactive' form added
in the previous commits.