magit-push: provide just one way to enable --set-upstream
The `magit-push-popup' features the `--set-upstream' argument, which is
how this sort of thing is usually done in Magit.  Until now `magit-push'
might also have asked the user whether she wants the upstream to be set.
The option `magit-set-upstream-on-push' was used to control under which
conditions the user was asked and the use of two prefix arguments did
override the option.  Remove the option and ignore prefix arguments with
regards to this.

Most things that was possible before are still possible, but there are
no longer multiple ways of doing the same thing.  The only exception is
that it is no longer possible to tell `magit-push' to refuse to push
when the upstream is not set.