add new push and pull commands
Previously only one command to push a branch and one command to pull a
branch existed.  Now there are several variants for pushing as well as
pulling.  They are all available from the respective popup which makes
it much easier to push to or pull from a branch which is not the already
configured upstream branch of the currently checked out branch.

Previously one had to remember exactly how many prefix arguments on had
to use to do something "unusual" and even then most "unusual" things
were not possible.  I for one did at least 2/3 of my pushes and pulls
in a terminal, but no more.

Add new commands `magit-pull-current', `magit-push-current',
`magit-push-elsewhere' and `magit-push-matching'.  Add the functions
`magit-pull-read-args' and `magit-push-read-args.  And adjust the
`magit-pull-popup' and `magit-push-popup popups, as well as the
function `magit-read-remote-branch' accordingly.