magit-diff-highlight-lines: keep foreground of other lines
By default only change the background color of lines that are outside
the region selecting lines in a hunk section.  Faces for added/removed
to use distinct colors for the background as well as the foreground, by
only changing the background to be identical to that of context lines,
added/removed lines outside the region look different from context
lines, and at the same time look different from selected added/removed
lines.

Add the variable `magit-diff-unmarked-lines-keep-foreground' setting
it to nil restores the old look.

Closes #1583.