magit-log-popup: bind lowercase keys to arguments, most actions
For keys that change the values of arguments, give up on only using
keys corresponding to the short form of the respective Git argument.
Instead use all lowercase letters and try to make them consistent
within this popup.

For action keys we cannot do completely without uppercase letters,
we use uppercase for the "verbose" variants.  However use "o"/"O"
instead of "b"/"B", for consistency with other popups where we also
use the former for the "act on something other than the default"
action variants.