magit-log-nongraph-args: new variable
Add new variable `magit-log-nongraph-args' which lists the arguments
that cannot be used together with `--graph' (doing so would result in
a broken and effectively meaningless graph).  We did already filter
out `--graph' before, but the incompatible arguments were hard-coded.

Perform the filtering in `magit-log', previously it was done in
`magit-log-read-args'.