improve magit-show-refs variants
Add `magit-show-refs-current', like `magit-show-refs-head' previously
did it compares the other refs with that branch.  The latter now
compares the other refs with `HEAD'.  The difference is that the
"current" variant locks the refs buffer to a branch; after checking
out another branch and refreshing the buffer, it continues to compare
the other refs with the (previously current) branch, while the "head"
variant locks to the `HEAD', checking out another branch or detaching
`HEAD' will cause the refs to be compared with another rev, the one
`HEAD' then points to.

Low level support for this distinction was implemented quite some
time ago but the commands did not take advantage of this distinction.
The only thing that changes is that level is that "HEAD" now stands
for itself, previously the "head variant" was represented by `nil'.