use saved arguments when command is invoked outside popup
When a command which is usually invoked as a popup action is invoked
directly then use the global value of the respective `PREFIX-arguments'
variable.

`magit-define-popup' now also defines a function by the same name.
When that function is called while a action is being invoked from the
respective popup, then it returns `magit-current-prefix-args's value.
Otherwise it returns the value of the variable `PREFIX-arguments'.

Replace most occurrences of `magit-current-prefix-args' with a call to
such a function.