magit-log-file: always show visited file, remove from popup
Remove `magit-log-file' from the `magit-log-popup' and stop reading
a file from the user and instead unconditionally show the log for
the visited file.  That's its only purpose now, to make that more
obvious rename it to `magit-log-buffer-file'.

The log commands that remain in the log popup will soon be taught how
to limit the history to a file or multiple files, while keeping their
other features which `magit-log-file' lacked.