magit-log{,-current,-head}: allow limiting to files
Teach log commands that are commonly invoked from the log popup to
limit history to commits which touch certain files selected in the
popup using the new "-- " argument.

Add `magit-log-read-files', which uses `completing-read-multiple', to
read multiple files from the user.  Store the read value as a single
string, because that's all that the current popup implementation
supports.  Then in `magit-log-read-args' separate that string from
the other arguments, split it into a list of files again, and return
that list separate from the other arguments.

Adjust `magit-log-refresh-buffer', `magit-insert-log` and
`magit-insert-log-verbose' accordingly.

Remove `--relative' from `magit-log-popup', it's no longer required and
confusing to non-experts.