magit-insert-untracked-files: respect status.showUntrackedFiles
Showing all untracked files can be slow.  So respect the Git variable
`status.showUntrackedFiles' allowing users to make it faster at the
cost of not being able to stage every untracked file individually.

When the variable is unset `git status' defaults to `normal' while
`magit-insert-untracked-files' defaults to `all', allowing them to
be staged individually.

Note that when `normal' is used the untracked files are still displayed
as a tree, but some of the trees cannot be expanded.  If directory
`mixed' contains an untracked and a tracked file then it is expandable
and expanding it will show the untracked file.  While the directory
`untracked' contains only untracked files and cannot be expanded.  With
everything expanded this would look like this:

Untracked files (2)
mixed/
mixed/untracked
untracked/