magit-diff-visit-file-worktree: new command
Jump to the file at point attempting (and usually failing) to also jump
to the correct line.  This is what plain `magit-diff-visit-file' used
to do until a while ago; now that jumps to the correct location in the
correct revision, but sometimes that's not what we want.

Bind C-return to the new command, so `magit-dired-jump' has to move to
M-return.