magit-branch-and-checkout: also starting from a stash
This uses `git stash branch' internally, which creates a branch, checks
it out, and then even modifies the index and/or worktree.  The last part
can be a little confusing for someone using `magit-branch-and-checkout'
with "stash@{N}" as starting point for the first time.  It would be even
more confusing when regular `magit-branch' ("and not checkout") did it,
so don't.