Using magit-run-git-async for stashing
This solves issue #1647 for git stash communicating with GPG. My guess
is that the asynchronous process is actually able to read the tty that
it needs to talk with gpg-agent.

I looked at the same code on the 'next' branch, and `magit-stash` has
been totally redone. I was unable to properly reproduce my issue on that
branch, but that's because it was not working in other, unrelated, ways.

That said, this is more of a hotfix since it is branched off 'master'.