magit-apply: don't always use --3way
With this argument changes are applied to the index in addition to
the working tree, which usually is not desirable.  And if the file
is question modified in the working tree, then it will cause `git
apply' to always fail.

With a prefix argument still use `--3way', but remove the option
`magit-apply-tree-way'.