magit-revert-buffers: refactor for speed and sanity
Use `magit-toplevel-safe' instead of `magit-get-top-dir'.  The latter
returns non-nil in bare repositories also, which is not helpful because
we only refresh files in the working directory anyway.

Use `file-in-directory-p' instead of `string-prefix-p', because the
former is safer.  Only check that the file is in the repository; do not
also check that it is _not_ in the git directory.  That's unnecessary
because a git repository cannot track a file ".git/FILE" anyway, and we
already check whether the file is a member of the list of tracked files.

Note that `magit-toplevel' and by extension `magit-toplevel-safe' use
`git rev-parse --show-level' which chases symlinks.  So we have to take
care when using `file-in-directory-p' or similar: we have to compare the
`buffer-file-truename' and not `buffer-file-name'.  We already did that.

Depending on whether there are more tracked files in the current
repositories or open buffers overall, take a different code path,
avoiding unnecessary work.  Move some of the shared code into the new
function `magit-revert-buffer'.