sanitize after-revert hooks
Rename the hook `magit-revert-buffer-hook' to `magit-after-revert-hook',
for consistency with the built-in `after-revert-hook'.  Run the hook,
after potentially re-enabling `magit-blame-mode' (which is a special
case explicitly handled in `magit-revert-buffer'), previously we did
it before.

Add the new hook variable `magit-not-reverted-hook' which is
run by `magit-revert-buffer' if, and only if, it does *not* run
`magit-after-revert-hook' (because there was nothing to be reverted,
i.e. the file did not change on disk).  This is for the benefit of
third-party extensions which do need to do something in that case.

Finish by always calling the _function_ `vc-find-file-hook'.  This
function is intended as a member of the hook `find-file-hook'.  It
is also called directly in a few other places, this is one of them.
Previously this function was only called when the buffer had to be
refreshed, now it is always called, mostly so that the mode-line
get's updated (that might need updating even if the file did not
change).