add smerge-refine-ignore-whitespace to magit-diff Custom group
Because it is defined as a regular variable, not a customizable option
belonging to some other group, it shows up as "rogue".  That's okay,
it's more important that users note that this can be customized than it
is that nothing is fishy about how it is presented in the Custom buffer.